LPS-45461 Create a new SettingsDefaultConfigurationAction class for dealing with Settings instead of PortletPreferences.
At the same time, the common functionality between DefaultConfigurationAction and SettingsDefaultConfigurationAction is moved into a common superclass called BaseDefaultConfigurationAction.
