+++ portal-service/src/com/liferay/portal/kernel/portlet/DefaultConfigurationAction.java
@@ -14,361 +14,81 @@
 
 package com.liferay.portal.kernel.portlet;
 
+import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.exception.SystemException;
-import com.liferay.portal.kernel.servlet.SessionErrors;
-import com.liferay.portal.kernel.servlet.SessionMessages;
-import com.liferay.portal.kernel.util.Constants;
-import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.LocaleUtil;
-import com.liferay.portal.kernel.util.LocalizationUtil;
-import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.portal.kernel.util.PropertiesParamUtil;
-import com.liferay.portal.kernel.util.StringPool;
-import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.kernel.util.UnicodeProperties;
-import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.kernel.util.WebKeys;
-import com.liferay.portal.model.Layout;
-import com.liferay.portal.model.Portlet;
-import com.liferay.portal.security.permission.ActionKeys;
-import com.liferay.portal.service.PortletLocalServiceUtil;
-import com.liferay.portal.service.permission.PortletPermissionUtil;
-import com.liferay.portal.settings.Settings;
-import com.liferay.portal.settings.SettingsFactoryUtil;
-import com.liferay.portal.theme.ThemeDisplay;
-import com.liferay.portal.util.PortalUtil;
-import com.liferay.portlet.PortletConfigFactoryUtil;
 
-import java.util.HashMap;
-import java.util.Map;
+import java.io.IOException;
 
 import javax.portlet.ActionRequest;
-import javax.portlet.ActionResponse;
-import javax.portlet.PortletConfig;
 import javax.portlet.PortletPreferences;
 import javax.portlet.PortletRequest;
-import javax.portlet.RenderRequest;
-import javax.portlet.RenderResponse;
-import javax.portlet.ResourceRequest;
-import javax.portlet.ResourceResponse;
+import javax.portlet.ReadOnlyException;
 import javax.portlet.ValidatorException;
 
-import javax.servlet.ServletContext;
-
 /**
- * @author Brian Wing Shun Chan
- * @author Julio Camarero
- * @author Raymond Augé
+ * @author Iván Zaera
  */
 public class DefaultConfigurationAction
-	extends LiferayPortlet
+	extends BaseDefaultConfigurationAction<PortletPreferences>
 	implements ConfigurationAction, ResourceServingConfigurationAction {
 
 	public static final String PREFERENCES_PREFIX = "preferences--";
 
-	public String getLocalizedParameter(
-		PortletRequest portletRequest, String name) {
-
-		String languageId = ParamUtil.getString(portletRequest, "languageId");
-
-		return getLocalizedParameter(portletRequest, name, languageId);
+	public DefaultConfigurationAction() {
+		super(PREFERENCES_PREFIX);
 	}
 
-	public String getLocalizedParameter(
-		PortletRequest portletRequest, String name, String languageId) {
-
-		return getParameter(
-			portletRequest,
-			LocalizationUtil.getLocalizedName(name, languageId));
+	@Override
+	protected PortletPreferences getConfiguration(ActionRequest actionRequest) {
+		return actionRequest.getPreferences();
 	}
 
-	public String getParameter(PortletRequest portletRequest, String name) {
-		name = PREFERENCES_PREFIX.concat(name).concat(StringPool.DOUBLE_DASH);
-
-		return ParamUtil.getString(portletRequest, name);
-	}
-
-	public void postProcessPortletPreferences(
+	@Override
+	protected void postProcess(
 			long companyId, PortletRequest portletRequest,
 			PortletPreferences portletPreferences)
-		throws Exception {
+		throws PortalException, SystemException {
 	}
 
 	@Override
-	public void processAction(
-			PortletConfig portletConfig, ActionRequest actionRequest,
-			ActionResponse actionResponse)
-		throws Exception {
-
-		String cmd = ParamUtil.getString(actionRequest, Constants.CMD);
-
-		if (!cmd.equals(Constants.UPDATE)) {
-			return;
-		}
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		Layout layout = PortletConfigurationLayoutUtil.getLayout(themeDisplay);
-
-		String portletResource = ParamUtil.getString(
-			actionRequest, "portletResource");
-
-		PortletPermissionUtil.check(
-			themeDisplay.getPermissionChecker(), themeDisplay.getScopeGroupId(),
-			layout, portletResource, ActionKeys.CONFIGURATION);
-
-		PortletPreferences portletPreferences = null;
-		Settings settings = null;
-
-		String settingsScope = ParamUtil.getString(
-			actionRequest, "settingsScope");
-
-		UnicodeProperties properties = PropertiesParamUtil.getProperties(
-			actionRequest, PREFERENCES_PREFIX);
-
-		if (Validator.isNull(settingsScope)) {
-			portletPreferences = actionRequest.getPreferences();
-
-			for (Map.Entry<String, String> entry : properties.entrySet()) {
-				String name = entry.getKey();
-				String value = entry.getValue();
-
-				portletPreferences.setValue(name, value);
-			}
-
-			Map<String, String[]> portletPreferencesMap =
-				(Map<String, String[]>)actionRequest.getAttribute(
-					WebKeys.PORTLET_PREFERENCES_MAP);
-
-			if (portletPreferencesMap != null) {
-				for (Map.Entry<String, String[]> entry :
-						portletPreferencesMap.entrySet()) {
-
-					String name = entry.getKey();
-					String[] values = entry.getValue();
-
-					portletPreferences.setValues(name, values);
-				}
-			}
-
-			postProcessPortletPreferences(
-				themeDisplay.getCompanyId(), actionRequest, portletPreferences);
-		}
-		else {
-			String serviceName = ParamUtil.getString(
-				actionRequest, "serviceName");
-
-			if (settingsScope.equals("company")) {
-				settings = SettingsFactoryUtil.getCompanyServiceSettings(
-					themeDisplay.getCompanyId(), serviceName);
-			}
-			else if (settingsScope.equals("group")) {
-				settings = SettingsFactoryUtil.getGroupServiceSettings(
-					themeDisplay.getSiteGroupId(), serviceName);
-			}
-			else if (settingsScope.equals("portletInstance")) {
-				settings = SettingsFactoryUtil.getPortletInstanceSettings(
-					themeDisplay.getLayout(), portletResource);
-			}
-
-			if (settings != null) {
-				for (Map.Entry<String, String> entry : properties.entrySet()) {
-					String name = entry.getKey();
-					String value = entry.getValue();
-
-					settings.setValue(name, value);
-				}
-
-				Map<String, String[]> portletPreferencesMap =
-					(Map<String, String[]>)actionRequest.getAttribute(
-						WebKeys.PORTLET_PREFERENCES_MAP);
-
-				if (portletPreferencesMap != null) {
-					for (Map.Entry<String, String[]> entry :
-							portletPreferencesMap.entrySet()) {
-
-						String name = entry.getKey();
-						String[] values = entry.getValue();
-
-						settings.setValues(name, values);
-					}
-				}
-			}
-		}
-
-		if (SessionErrors.isEmpty(actionRequest)) {
-			try {
-				if (portletPreferences != null) {
-					portletPreferences.store();
-				}
-
-				if (settings != null) {
-					settings.store();
-				}
-			}
-			catch (ValidatorException ve) {
-				SessionErrors.add(
-					actionRequest, ValidatorException.class.getName(), ve);
-
-				return;
-			}
-
-			SessionMessages.add(
-				actionRequest,
-				PortalUtil.getPortletId(actionRequest) +
-					SessionMessages.KEY_SUFFIX_REFRESH_PORTLET,
-				portletResource);
-
-			SessionMessages.add(
-				actionRequest,
-				PortalUtil.getPortletId(actionRequest) +
-					SessionMessages.KEY_SUFFIX_UPDATED_CONFIGURATION);
-
-			String redirect = PortalUtil.escapeRedirect(
-				ParamUtil.getString(actionRequest, "redirect"));
-
-			if (Validator.isNotNull(redirect)) {
-				actionResponse.sendRedirect(redirect);
-			}
-		}
-	}
-
-	@Override
-	public String render(
-			PortletConfig portletConfig, RenderRequest renderRequest,
-			RenderResponse renderResponse)
-		throws Exception {
-
-		PortletConfig selPortletConfig = getSelPortletConfig(renderRequest);
-
-		String configTemplate = selPortletConfig.getInitParameter(
-			"config-template");
-
-		if (Validator.isNotNull(configTemplate)) {
-			return configTemplate;
-		}
-
-		String configJSP = selPortletConfig.getInitParameter("config-jsp");
-
-		if (Validator.isNotNull(configJSP)) {
-			return configJSP;
-		}
-
-		return "/configuration.jsp";
-	}
-
-	@Override
-	public void serveResource(
-			PortletConfig portletConfig, ResourceRequest resourceRequest,
-			ResourceResponse resourceResponse)
-		throws Exception {
-	}
-
-	public void setPreference(
-		PortletRequest portletRequest, String name, String value) {
-
-		setPreference(portletRequest, name, new String[] {value});
-	}
-
-	public void setPreference(
-		PortletRequest portletRequest, String name, String[] values) {
-
-		Map<String, String[]> portletPreferencesMap =
-			(Map<String, String[]>)portletRequest.getAttribute(
-				WebKeys.PORTLET_PREFERENCES_MAP);
-
-		if (portletPreferencesMap == null) {
-			portletPreferencesMap = new HashMap<String, String[]>();
-
-			portletRequest.setAttribute(
-				WebKeys.PORTLET_PREFERENCES_MAP, portletPreferencesMap);
-		}
-
-		portletPreferencesMap.put(name, values);
-	}
-
-	protected PortletConfig getSelPortletConfig(PortletRequest portletRequest)
-		throws SystemException {
-
-		ThemeDisplay themeDisplay = (ThemeDisplay)portletRequest.getAttribute(
-			WebKeys.THEME_DISPLAY);
-
-		String portletResource = ParamUtil.getString(
-			portletRequest, "portletResource");
-
-		Portlet selPortlet = PortletLocalServiceUtil.getPortletById(
-			themeDisplay.getCompanyId(), portletResource);
-
-		ServletContext servletContext =
-			(ServletContext)portletRequest.getAttribute(WebKeys.CTX);
-
-		PortletConfig selPortletConfig = PortletConfigFactoryUtil.create(
-			selPortlet, servletContext);
-
-		return selPortletConfig;
-	}
-
-	protected void removeDefaultValue(
-			PortletRequest portletRequest,
-			PortletPreferences portletPreferences, String key,
-			String defaultValue)
-		throws Exception {
-
-		String value = getParameter(portletRequest, key);
-
-		if (defaultValue.equals(value) ||
-			StringUtil.equalsIgnoreBreakLine(defaultValue, value)) {
-
+	protected void reset(PortletPreferences portletPreferences, String key) {
+		try {
 			portletPreferences.reset(key);
 		}
-	}
-
-	protected void validateEmail(
-		ActionRequest actionRequest, String emailParam, boolean localized) {
-
-		boolean emailEnabled = GetterUtil.getBoolean(
-			getParameter(actionRequest, emailParam + "Enabled"));
-		String emailSubject = null;
-		String emailBody = null;
-
-		if (localized) {
-			String languageId = LocaleUtil.toLanguageId(
-				LocaleUtil.getSiteDefault());
-
-			emailSubject = getLocalizedParameter(
-				actionRequest, emailParam + "Subject", languageId);
-			emailBody = getLocalizedParameter(
-				actionRequest, emailParam + "Body", languageId);
-		}
-		else {
-			emailSubject = getParameter(actionRequest, emailParam + "Subject");
-			emailBody = getParameter(actionRequest, emailParam + "Body");
-		}
-
-		if (emailEnabled) {
-			if (Validator.isNull(emailSubject)) {
-				SessionErrors.add(actionRequest, emailParam + "Subject");
-			}
-			else if (Validator.isNull(emailBody)) {
-				SessionErrors.add(actionRequest, emailParam + "Body");
-			}
+		catch (ReadOnlyException roe) {
+			throw new RuntimeException(roe);
 		}
 	}
 
-	protected void validateEmailFrom(ActionRequest actionRequest) {
-		String emailFromName = getParameter(actionRequest, "emailFromName");
-		String emailFromAddress = getParameter(
-			actionRequest, "emailFromAddress");
+	@Override
+	protected void setValue(
+		PortletPreferences portletPreferences, String name, String value) {
 
-		if (Validator.isNull(emailFromName)) {
-			SessionErrors.add(actionRequest, "emailFromName");
+		try {
+			portletPreferences.setValue(name, value);
 		}
-		else if (!Validator.isEmailAddress(emailFromAddress) &&
-				 !Validator.isVariableTerm(emailFromAddress)) {
+		catch (ReadOnlyException roe) {
+			throw new RuntimeException(roe);
+		}
+	}
 
-			SessionErrors.add(actionRequest, "emailFromAddress");
+	@Override
+	protected void setValues(
+		PortletPreferences portletPreferences, String name, String[] values) {
+
+		try {
+			portletPreferences.setValues(name, values);
 		}
+		catch (ReadOnlyException roe) {
+			throw new RuntimeException(roe);
+		}
+	}
+
+	@Override
+	protected void store(PortletPreferences portletPreferences)
+		throws IOException, ValidatorException {
+
+		portletPreferences.store();
 	}
 
 }
